Improve performance of entropy calculation
This repo is the 4th google result for searches like "python calculate entropy of file" and the 1st github repo. It is unfortunately slow. The currently implementation scans the file 256 times to check each byte value. This is not necessary if we keep track of the byte as we see them.  This could be done with a `dict`, however I have chosen to use `collections.Counter` which is a subclass of `dict`.

I have also removed unnecessary casts and restructured one of the for loops as a list comprehension, this avoids calls to `list().append()` and uses special opcodes to loop faster.